Enforce existence of CashCard and valid ownership

+++ src/main/java/com/waihon/cashcard/controller/CashCardController.java
@@ -2,7 +2,6 @@ package com.waihon.cashcard.controller;
 
 import com.waihon.cashcard.entity.CashCard;
 import com.waihon.cashcard.repository.CashCardRepository;
-import org.apache.coyote.Response;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -105,7 +104,10 @@ class CashCardController {
     }
 
     @DeleteMapping("/{id}")
-    private ResponseEntity<Void> deleteCashCard(@PathVariable Long id) {
+    private ResponseEntity<Void> deleteCashCard(@PathVariable Long id, Principal principal) {
+        if (!cashCardRepository.existsByIdAndOwner(id, principal.getName())) {
+            return ResponseEntity.notFound().build();
+        }
         cashCardRepository.deleteById(id);
         return ResponseEntity.noContent().build();
     }
